requirements.txt
langchain
langgraph
langchain-openai
python-dotenv
langchain_google_genai
streamlit
langgraph-checkpoint-sqlite
psycopg[binary,pool]
langgraph-checkpoint-postgres
langchain_community
ddgs